pyproject.toml
[project]
name = "demo2"
version = "0.1.1"
description = "A Pygame example project"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pygame-ce>=2.5.2",
    "pygbag>=0.9.2",
]
